requirements.txt
asgiref==3.10.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==2.2.0
Django==5.1.4
django-celery-beat==2.8.1
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
redis==7.0.0
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
